[MAC] WIP(808): Initial attempt at changing cursor over hyperlinks
Apparently because Keyman never becomes the active app (even when it is displaying configuration windows that are key windows), all attempts to get the cursor to change fail. Presumably that's why it's not happening automatically in the first place. In this commit, I was able to add a label that shows the destination URL (not quite correct, since we'd probably want to trim off the keyman?link:url= prefix for links that ae targeting the default browser) when the mouse passes over them. That's probably better than noting.